refactor(agent): isolate streaming wait monitor phase
